+++ Src/AppFramework/Abstractions/AppAdType.cs
@@ -1,0 +1,8 @@
+﻿namespace VpnHood.AppFramework.Abstractions;
+
+public enum AppAdType
+{
+    InterstitialAd,
+    RewardedAd,
+    AppOpenAd
+}