+++ Src/AppFramework/Abstractions/IAppAdProvider.cs
@@ -1,0 +1,13 @@
+﻿using VpnHood.Client.Device;
+
+namespace VpnHood.AppFramework.Abstractions;
+
+public interface IAppAdProvider : IDisposable
+{
+    string NetworkName { get; }
+    AppAdType AdType { get; }
+    Task LoadAd(IUiContext uiContext, CancellationToken cancellationToken);
+    Task ShowAd(IUiContext uiContext, string? customData, CancellationToken cancellationToken);
+    DateTime? AdLoadedTime { get; }
+    TimeSpan AdLifeSpan { get; }
+}