+++ Src/AppFramework/Abstractions/IAppUpdaterProvider.cs
@@ -1,0 +1,8 @@
+﻿using VpnHood.Client.Device;
+
+namespace VpnHood.AppFramework.Abstractions;
+
+public interface IAppUpdaterProvider
+{
+    Task<bool> Update(IUiContext uiContext);
+}