+++ Src/AppFramework/AppFramework/ClientProfiles/ClientProfileExtensions.cs
@@ -1,0 +1,23 @@
+﻿namespace VpnHood.AppFramework.ClientProfiles;
+
+public static class ClientProfileExtensions
+{
+    public static ClientProfileInfo ToInfo(this ClientProfile clientProfile)
+    {
+        return new ClientProfileInfo(clientProfile);
+    }
+
+    public static ClientProfileBaseInfo ToBaseInfo(this ClientProfileInfo clientProfileInfo)
+    {
+        return new ClientProfileBaseInfo {
+            ClientProfileId = clientProfileInfo.ClientProfileId,
+            ClientProfileName = clientProfileInfo.ClientProfileName,
+            SupportId = clientProfileInfo.SupportId,
+            CustomData = clientProfileInfo.CustomData,
+            IsPremiumLocationSelected = clientProfileInfo.IsPremiumLocationSelected,
+            IsPremiumAccount = clientProfileInfo.IsPremiumAccount,
+            SelectedLocationInfo = clientProfileInfo.SelectedLocationInfo,
+            HasAccessCode = !string.IsNullOrEmpty(clientProfileInfo.AccessCode)
+        };
+    }
+}