+++ Src/AppFramework/Win.Common.WpfSpa/WinUiContext.cs
@@ -1,0 +1,9 @@
+﻿using System.Windows;
+using VpnHood.Client.Device;
+
+namespace VpnHood.AppFramework.Win.Common.WpfSpa;
+
+public class WinUiContext(Window window) : IUiContext
+{
+    public bool IsActive => window.IsActive;
+}