+++ VpnHood.Client.App.Libs/VpnHood.Client.App.Abstractions/IAppUpdaterProvider.cs
@@ -1,8 +1,0 @@
-﻿using VpnHood.Client.Device;
-
-namespace VpnHood.Client.App.Abstractions;
-
-public interface IAppUpdaterProvider
-{
-    Task<bool> Update(IUiContext uiContext);
-}