+++ CHANGELOG.md
@@ -898,7 +898,7 @@
 ### Client
 * Update: Add client prefix to Bug Report File Name
 * Update: Close Bug Report bottom page after sending report
-* Update: Separate SPA from VpnHood.Client.App.UI. Make it easier for developers to use custom SPA
+* Update: Separate SPA from VpnHood.AppFramework.UI. Make it easier for developers to use custom SPA
 * Update: Change Anonymous IP masking from *.*.x.x to "*.x.x.*"
 * Update: Diagnose set Last error to "Diagnose has been finished" if there is not other error
 * Fix: Dark Icon